requirements.txt
flask==3.1.3
groq==1.1.2
networkx==3.6.1
pyvis==0.3.2
python-dotenv==1.2.2
gunicorn==21.2.0